macOS extension localization improvement
